Integrate utilization of log4cxx-based logging.
Use of the new logging is *optionally* used within ccontrol module
(AsyncQueryManager.cc, dispatcher.cc, and thread.cc) as well as the
Python layer.

A preprocessor #define (NEWLOG) determines whether or not the new
logging is used within the ccontrol module. Meanwhile, a Python
module-scoped boolean variable (also NEWLOG) within logger.py of the
czar module toggles the new logging on/off for the Python-layer.

It's important to note that these changes are incomplete and have not
been tested. They are meant only to serve as an examples of how to
incorporate the new logging solution into Qserv. To complete
integeration of the new logging solution, the new log4cxx-based log
system needs to be packaged (via eups) and included as part of the
standard Qserv distribution.