Copy previous reject-non-dummy-chunks-for-non-partitioned-queries feature.
The c++ level rejects adding chunks for non-partitioned queries so as to
relieve the Python layer of worrying about it (this is one step in pushing
query coverage control to c++ rather than python). This rejection wasn't
ported when switching the API to the UserQuery abstraction. This commit
fixes this.