Implement refreshing of CSS cache.
Add support for refreshing CSS cache. Refresh frequency can be specified through
a config file. If not specified, default is 15 sec. If a query or multiple queries)
are still using Facade, it will be kept until it is no longer used. That means
that there might be multiple Facade caches kept. New queries always get the most
recent cache. Older caches are automatically purged when they are no longer used.