Implement returning results of disconnected queries/jobs (DM-11191)
New UserQuery subclass (UserQueryAsyncResult) handles most of the job.
It finds necessary info in QMeta and when query is finished it pretends
that its own result table is the result table of original query. Message
table of original query needs to be copied to a new message table.
Presently result can be returned only once (proxy drops result table),
in the future we'll need better lifetime management for results.